Add authentication, CRUD improvements, loading states and 404 page

+++ src/components/EmployeesPage.vue
@@ -1,33 +1,25 @@
 <template>
     <NavBar />
     <!-- <h3>Employees Record</h3> -->
-     <div class="container">
+    <div class="container">
+        <p class="alert alert-danger" v-if="error">{{ error }}</p>
         <div class="row mb-3">
             <div class="col-3">
-                <input 
-                    type="text" 
-                    class="form-control" 
-                    placeholder="Search by name" 
-                    v-model="firstName"
-                    >
-            </div>
-             <div class="col-3">
-                <input 
-                    type="text" class="form-control" 
-                    placeholder="Search by email" 
-                    v-model="email">
+                <input type="text" class="form-control" placeholder="Search by name" v-model="firstName">
             </div>
             <div class="col-3">
-                <input 
-                    type="text" 
-                    class="form-control" 
-                    placeholder="Search by gender" 
-                    v-model="gender">
+                <input type="text" class="form-control" placeholder="Search by email" v-model="email">
+            </div>
+            <div class="col-3">
+                <input type="text" class="form-control" placeholder="Search by gender" v-model="gender">
             </div>
         </div>
-     </div>
-    <div class="container-fluid">
-        <table class="table table-bordered" >
+    </div>
+    <div v-if="filteredUsers.length === 0" class="alert alert-info text-center">
+        No employees found.
+    </div>
+    <div v-else class="container-fluid">
+        <table class="table table-bordered">
             <thead>
                 <tr>
                     <th>Name</th>
@@ -44,12 +36,12 @@
                     <td>{{ user.gender }}</td>
                     <td>{{ user.age }}</td>
                     <td>
-                <router-link class="btn" :to="`/edit-employee/${user.id}`">
-                    Edit
-                </router-link>     
-                  <button @click="deleteEmployee(user.id)" class="btn btn-primary">
-                        Delete
-                    </button>               
+                        <router-link class="btn" :to="`/edit-employee/${user.id}`">
+                            Edit
+                        </router-link>
+                        <button @click="deleteEmployee(user.id)" class="btn btn-primary" :disabled="loading">
+                            {{ loading ? 'Deleting employee...' : 'Delete' }}
+                        </button>
                     </td>
                 </tr>
             </tbody>
@@ -63,53 +55,75 @@ import NavBar from './NavBar.vue';
 
 export default {
     name: 'EmployessPage',
-    components:{
+    components: {
         NavBar
     },
     data() {
         return {
             users: [],
-            firstName:'',
-            email:'',
-            gender:'',
-            
+            firstName: '',
+            email: '',
+            gender: '',
+            error: '',
+            loading: true,
+
         }
     },
-    computed:{
+    computed: {
         filteredUsers() {
             return this.users.filter(user => {
                 const searchName = user.firstName
-               .toLowerCase().includes(this.firstName.toLowerCase());
+                    .toLowerCase().includes(this.firstName.toLowerCase());
                 const searchEmail = user.email.
-                toLowerCase().includes(this.email.toLowerCase());
+                    toLowerCase().includes(this.email.toLowerCase());
                 const searchGender = user.gender
-                .toLowerCase().includes(this.gender.toLowerCase());
-               return searchName && searchEmail && searchGender
+                    .toLowerCase().includes(this.gender.toLowerCase());
+                return searchName && searchEmail && searchGender
 
             })
         }
     },
     async mounted() {
-        const result = await axios.get('https://dummyjson.com/users')
-        this.users = result.data.users.slice(0, 10)
-        console.log(result);
-        console.log(result.data);
-        console.log(result.data.users);
+        try {
+            this.loading = true;
+            const result = await axios.get('https://dummyjson.com/users')
+            this.users = result.data.users.slice(0, 10)
+            console.log(result);
+            console.log(result.data);
+            console.log(result.data.users);
+            this.loading = false;
+
+        } catch (error) {
+            this.error = 'Failed to get data',
+                this.loading = false;
+        }
 
     },
     methods: {
         async deleteEmployee(id) {
-            const result = await axios.delete(`https://dummyjson.com/users/${id}`)
-            this.users = this.users.filter(user => user.id !== id);
-            console.log(this.users)
-            console.log(result.data);
+            try {
+                const confirmResult = confirm('are you sure you want to delete this employee')
+                if (confirmResult) {
+                    this.loading = true;
+                    const result = await axios.delete(`https://dummyjson.com/users/${id}`)
+                    this.users = this.users.filter(user => user.id !== id);
+                    console.log(this.users)
+                    console.log(result.data);
+                }
+                this.loading = false;
+            } catch (error) {
+                this.error = 'Failed to delete employee',
+                    this.loading = false;
+                // this.$router.push('/')
+            }
         }
     }
 }
 </script>
 
 <style>
-.btn, .btn:hover {
+.btn,
+.btn:hover {
     background-color: darkblue;
     color: white;
     margin: 2px;
